rapido_system/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
asttokens==3.0.0
attrs==25.3.0
audioread==3.0.1
av==15.1.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
cupy-cuda11x==13.6.0
decorator==5.2.1
docstring_parser==0.17.0
elevenlabs==2.15.0
eval_type_backport==0.2.2
executing==2.2.1
fastapi==0.116.1
fastrlock==0.8.3
ffmpeg-python==0.2.0
filelock==3.13.1
frozenlist==1.7.0
fsspec==2024.6.1
future==1.0.0
googleapis-common-protos==1.70.0
GPUtil==1.4.0
greenlet==3.2.4
grpcio==1.74.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.10
importlib_metadata==8.7.0
ipython==9.5.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.4
joblib==1.5.2
json5==0.9.14
lazy_loader==0.4
librosa==0.10.1
livekit==1.0.12
livekit-agents==1.2.8
livekit-api==1.0.5
livekit-blingfire==1.0.0
livekit-plugins-elevenlabs==1.2.8
livekit-protocol==1.0.5
llvmlite==0.44.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mpmath==1.3.0
msgpack==1.1.1
multidict==6.6.4
nest-asyncio==1.6.0
networkx==3.3
numba==0.61.2
numpy==2.2.6
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.1.105
opencv-contrib-python==4.12.0.88
opencv-python==4.8.1.78
opentelemetry-api==1.36.0
opentelemetry-exporter-otlp==1.36.0
opentelemetry-exporter-otlp-proto-common==1.36.0
opentelemetry-exporter-otlp-proto-grpc==1.36.0
opentelemetry-exporter-otlp-proto-http==1.36.0
opentelemetry-proto==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
packaging==25.0
parso==0.8.5
pexpect==4.9.0
Pillow==10.0.1
platformdirs==4.4.0
playwright==1.55.0
pooch==1.8.2
prometheus_client==0.22.1
prompt_toolkit==3.0.52
propcache==0.3.2
protobuf==6.32.0
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pyee==13.0.0
Pygments==2.19.2
PyJWT==2.10.1
python-dotenv==1.0.0
requests==2.31.0
scikit-learn==1.7.1
scipy==1.16.1
setuptools==70.2.0
sniffio==1.3.1
sounddevice==0.5.2
soundfile==0.12.1
soxr==1.0.0
stack-data==0.6.3
starlette==0.47.3
structlog==25.4.0
sympy==1.13.1
threadpoolctl==3.6.0
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
traitlets==5.14.3
triton==3.1.0
types-protobuf==6.30.2.20250822
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.35.0
watchfiles==1.1.0
wcwidth==0.2.13
websockets==11.0.3
yarl==1.20.1
zipp==3.23.0